+++ app/src/main/java/com/inky/fitnesscalendar/broadcast_receiver/NotificationBroadcastReceiver.kt
@@ -4,7 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import com.inky.fitnesscalendar.repository.DatabaseRepository
+import com.inky.fitnesscalendar.repository.RecordingRepository
 import com.inky.fitnesscalendar.util.ACTION_CANCEL
 import com.inky.fitnesscalendar.util.ACTION_SAVE
 import com.inky.fitnesscalendar.util.EXTRA_RECORDING_ID
@@ -19,7 +19,7 @@ private const val TAG = "NotificationPodcastReceiver"
 @AndroidEntryPoint
 class NotificationBroadcastReceiver : BroadcastReceiver() {
     @Inject
-    lateinit var repository: DatabaseRepository
+    lateinit var repository: RecordingRepository
 
     override fun onReceive(context: Context?, intent: Intent?) {
         Log.d(TAG, intent?.toUri(Intent.URI_INTENT_SCHEME) ?: "Intent was null")